Move sm2 ctx functions to a standalone source file
To support UADK implementations
